requirements.txt
selenium==4.10.0
beautifulsoup4==4.12.2
flask==2.3.3
pillow==10.1.0
werkzeug==2.3.7
jinja2==3.1.2
markupsafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
urllib3==2.0.7
webdriver-manager==4.0.1 